refactor(deps): ban all relative imports, fix remaining violations
Broaden no-relative-import rule to catch all relative imports (not just
.js). Migrate remaining relative imports in test files, test262-main
generators, and test-d files. Fix trailing-slash and double-slash import
paths.

Ignores (with reasons):
- scripts/tools: Node --experimental-transform-types can't resolve #packages/*.js to .ts
- *.test-d.ts: tsd uses separate tsconfig without #packages/ path mapping
- packer.ts: shared between src and scripts
- eslint-plugin-formatjs/index.ts: imports ./package.json (JSON)
- examples: separate ts_project
