+++ packages/intl-datetimeformat/test262-main.ts
@@ -1,7 +1,7 @@
 /* @generated */
 // @ts-nocheck
-import './polyfill-force.js'
-import allData from './data/all-tz.generated.ts'
+import '#packages/intl-datetimeformat/polyfill-force.js'
+import allData from '#packages/intl-datetimeformat/data/all-tz.generated.js'
 defineProperty(Intl, 'DateTimeFormat', {value: DateTimeFormat})
 Intl.DateTimeFormat.__addLocaleData(
   {
