+++ packages/intl-listformat/scripts/test262-main-gen.ts
@@ -17,7 +17,7 @@ function main(args: Args) {
     out,
     `// @generated
 // @ts-nocheck
-import './polyfill-force.js';
+import '#packages/intl-listformat/polyfill-force.js';
 if (Intl.ListFormat && typeof Intl.ListFormat.__addLocaleData === 'function') {
   Intl.ListFormat.__addLocaleData(${allData.join(',\n')})
 }`
