+++ packages/intl-locale/tests/index.test.ts
@@ -1,5 +1,5 @@
 import '@formatjs/intl-getcanonicallocales/polyfill.js'
-import {Locale} from '#packages/intl-locale/'
+import {Locale} from '#packages/intl-locale/index.js'
 import {describe, expect, it, test} from 'vitest'
 describe('intl-locale', () => {
   it('toString', function () {
