+++ packages/intl-numberformat/scripts/test262-main-gen.ts
@@ -17,7 +17,7 @@ function main(args: Args) {
     out,
     `// @generated
 // @ts-nocheck
-import './polyfill-force.ts';
+import '#packages/intl-numberformat/polyfill-force.js';
 if (Intl.NumberFormat && typeof Intl.NumberFormat.__addLocaleData === 'function') {
   Intl.NumberFormat.__addLocaleData(${allData.join(',\n')})
 }`
