+++ packages/intl-pluralrules/test262-main.ts
@@ -1,7 +1,7 @@
 /* @generated */
 // prettier-ignore
 // @ts-nocheck
-import './polyfill-force.ts'
+import '#packages/intl-pluralrules/polyfill-force.js'
 if (
   Intl.PluralRules &&
   typeof Intl.PluralRules.__addLocaleData === 'function'
