+++ packages/intl-pluralrules/tests/index.test.ts
@@ -1,6 +1,6 @@
 import '@formatjs/intl-getcanonicallocales/polyfill.js'
 import '@formatjs/intl-locale/polyfill.js'
-import {PluralRules} from '#packages/intl-pluralrules/'
+import {PluralRules} from '#packages/intl-pluralrules/index.js'
 import {describe, expect, it} from 'vitest'
 // @ts-ignore
 import en from '#packages/intl-pluralrules/tests/locale-data/en.js'
