+++ packages/intl-segmenter/scripts/test262-main-gen.ts
@@ -14,7 +14,7 @@ function main(args: Args) {
     out,
     `// @generated
 // @ts-nocheck
-import './polyfill-force.js';
+import '#packages/intl-segmenter/polyfill-force.js';
 `
   )
 }
