+++ packages/react-intl/tests/unit/types.test.tsx
@@ -1,4 +1,4 @@
-import {defineMessages, useIntl} from '#packages/react-intl/'
+import {defineMessages, useIntl} from '#packages/react-intl/index.js'
 import {describe, it} from 'vitest'
 describe('types', () => {
   // https://github.com/formatjs/formatjs/issues/3856
